+++ packages/lumx-icons/index.d.ts
@@ -1,3 +1,0 @@
-export * from '@mdi/js';
-
-export * from './v4-to-v5-aliases';