+++ packages/lumx-icons/override/override-icons/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    'google-pages': {
+        cssCode: '\\F2BA',
+    },
+};